crnn/requirements.txt
torch
torchvision
Pillow
tqdm
lmdb
opencv-python-headless
huggingface-hub
editdistance
numpy
